requirements.txt
pandas
python-docx
pdfplumber
reportlab
Pillow
markdown
html2text
openpyxl
moviepy
pydub
opencv-python
ffmpeg-python
PyQt5
PyYAML
beautifulsoup4
lxml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib